task(PR-Review): Review Comments
Basic addressing of comments and name changes. Refactor the routes file
to reflect the web version a bit more closely
