+++ src/navigation/index.js
@@ -1,16 +1,13 @@
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useContext, useEffect } from 'react';
 import { NativeModules } from 'react-native';
 import { NavigationContainer, DefaultTheme } from '@react-navigation/native';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+
 import Theme from '../theme/index';
-import { Login, Todos, Register } from '../screens';
-import Loading from '../components/utilities/loading';
 import { AppContext, useGlobalStateMgmt } from '../global-state';
 import { useToggle } from '../hooks/useToggle';
-import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
+import { LoginRoutes, TodoRoutes } from '../navigation/routes';
 
 const { ForgeRockModule } = NativeModules;
-const Tab = createBottomTabNavigator();
 
 function Navigation() {
   const [auth, setAuth] = useToggle();
@@ -32,7 +29,7 @@ function Navigation() {
     isAuthenticated: auth,
   });
 
-  [{ isAuthenticated }, { setAuthentication }] = stateMgmt;
+  const [{ isAuthenticated }] = stateMgmt;
 
   return (
     <Theme>
@@ -47,85 +44,10 @@ function Navigation() {
   );
 }
 
-function Logout() {
-  const [{}, { setAuthentication }] = useContext(AppContext);
-  useEffect(() => {
-    async function logout() {
-      await ForgeRockModule.performUserLogout();
-      setAuthentication(false);
-    }
-    logout();
-  }, []);
-
-  return <Loading />;
-}
-
 function RootNavigator() {
   const [{ isAuthenticated }] = useContext(AppContext);
 
-  return isAuthenticated ? (
-    <Tab.Navigator initialRoute="Todos">
-      <Tab.Screen
-        name="Todos"
-        component={Todos}
-        options={{
-          tabBarLabel: 'Todos',
-          tabBarIcon: ({ color, size }) => (
-            <MaterialCommunityIcons
-              name="format-list-checks"
-              color={color}
-              size={size}
-            />
-          ),
-        }}
-      />
-      <Tab.Screen
-        name="Logout"
-        component={Logout}
-        options={{
-          tabBarLabel: 'Logout',
-          tabBarIcon: ({ color, size }) => (
-            <MaterialCommunityIcons
-              name="logout-variant"
-              color={color}
-              size={size}
-            />
-          ),
-        }}
-      />
-    </Tab.Navigator>
-  ) : (
-    <Tab.Navigator initialRoute="Login">
-      <Tab.Screen
-        name="Login"
-        component={Login}
-        options={{
-          tabBarLabel: 'Login',
-          tabBarIcon: ({ color, size }) => (
-            <MaterialCommunityIcons
-              name="login-variant"
-              color={color}
-              size={size}
-            />
-          ),
-        }}
-      />
-      <Tab.Screen
-        name="Register"
-        component={Register}
-        options={{
-          tabBarLabel: 'Register',
-          tabBarIcon: ({ color, size }) => (
-            <MaterialCommunityIcons
-              name="account-plus"
-              color={color}
-              size={size}
-            />
-          ),
-        }}
-      />
-    </Tab.Navigator>
-  );
+  return isAuthenticated ? <TodoRoutes /> : <LoginRoutes />;
 }
 
 export default Navigation;
